+++ edsnlp/pipes/qualifiers/base.py
@@ -7,7 +7,11 @@ from spacy.tokens import Doc, Span
 from edsnlp.core import PipelineProtocol
 from edsnlp.matchers.phrase import EDSPhraseMatcher
 from edsnlp.matchers.regex import RegexMatcher
-from edsnlp.pipes.base import BaseComponent, SpanGetterArg, validate_span_getter
+from edsnlp.pipes.base import (
+    BaseSpanAttributeClassifierComponent,
+    SpanGetterArg,
+    validate_span_getter,
+)
 
 
 def check_normalizer(nlp: PipelineProtocol) -> None:
@@ -34,9 +38,10 @@ def get_qualifier_extensions(nlp: PipelineProtocol):
     }
 
 
-class RuleBasedQualifier(BaseComponent):
+class RuleBasedQualifier(BaseSpanAttributeClassifierComponent):
     """
-    Implements the NegEx algorithm.
+    Implements the ConText algorithm (eq. NegEx for negations) for detecting contextual
+    attributes text.
 
     Parameters
     ----------
@@ -69,14 +74,13 @@ class RuleBasedQualifier(BaseComponent):
         name: Optional[str] = None,
         *,
         attr: str,
+        attributes: List[str],
         span_getter: SpanGetterArg,
         on_ents_only: Union[bool, str, List[str], Set[str]],
         explain: bool,
         terms: Dict[str, Optional[List[str]]],
         regex: Dict[str, Optional[List[str]]] = {},
     ):
-        super().__init__(nlp=nlp, name=name)
-
         if attr.upper() == "NORM":
             check_normalizer(nlp)
 
@@ -85,6 +89,7 @@ class RuleBasedQualifier(BaseComponent):
 
         self.regex_matcher = RegexMatcher(attr=attr)
         self.regex_matcher.build_patterns(regex=regex)
+        self.attributes = attributes
 
         self.on_ents_only = on_ents_only
 
@@ -103,8 +108,12 @@ class RuleBasedQualifier(BaseComponent):
             span_getter = "ents" if on_ents_only is True else on_ents_only
         else:
             span_getter = "ents"
-        self.span_getter = validate_span_getter(span_getter)
         self.explain = explain
+        super().__init__(
+            nlp=nlp,
+            name=name,
+            span_getter=validate_span_getter(span_getter),
+        )
 
     def get_matches(self, doc: Doc) -> List[Span]:
         """